Move all db service files to dedicated services
