+++ modules/clubs/clubs.middleware.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler');
 const httpStatus = require('http-status');
 const Club = require('./club.model');
 const setAsset = require('../../middleware/setAsset');
-const dbService = require('../../services/db.service');
+const playersService = require('../players/players.service');
 const ApiError = require('../../utils/ApiError');
 
 const setClub = setAsset({ name: 'club', model: Club });
@@ -10,7 +10,7 @@ const setClub = setAsset({ name: 'club', model: Club });
 const canBeDeleted = asyncHandler(async (req, res, next) => {
   const { id } = req.params;
 
-  const players = await dbService.getPlayersForClub(id);
+  const players = await playersService.getPlayersForClub(id);
 
   if (players.length > 0) {
     return next(
