+++ modules/users/users.service.js
@@ -1,5 +1,12 @@
 const User = require('./user.model');
 
+function getUserById(id) {
+  return User.findById(id);
+}
+function getUserByConfirmationCode(confirmationCode) {
+  return User.findOne({ confirmationCode });
+}
+
 function createUser(reqBody) {
   return User.create(reqBody);
 }
@@ -7,13 +14,9 @@ function createUser(reqBody) {
 function getUserByEmail(email) {
   return User.findOne({ email });
 }
-
-function getUserByConfirmationCode(confirmationCode) {
-  return User.findOne({ confirmationCode });
-}
-
 module.exports = {
   createUser,
   getUserByEmail,
   getUserByConfirmationCode,
+  getUserById,
 };
